data/Non-AMP/requirements.txt
pandas>=1.3.0
numpy>=1.20.0
biopython>=1.79
matplotlib>=3.3.0
seaborn>=0.11.0
